copy code now has the ability to specify whether you want to delete the contents of the destination directory or not. Helpful for Node.js apps when only a code change is required. This will prevent the need to then install npm dependencies again

+++ bifrost/helpers/deploy.py
@@ -117,14 +117,16 @@ def convert_to_bool(val, default=False):
     except ValueError:
         return default
 
-def copy_code(deployment_config, source_dir='src', **additional_files):
+def copy_code(deployment_config, source_dir='src', delete_dest_contents=True,
+              **additional_files):
     """
     Removes the old code base before copying the new one across
     """
     deployment_dir = os.path.join(deployment_config['base_dir'],
                                   deployment_config['code_dir'].strip())
 
-    sudo('rm -rf %s/*' % deployment_dir, user=deployment_config['user'])
+    if delete_dest_contents:
+        sudo('rm -rf %s/*' % deployment_dir, user=deployment_config['user'])
     sudo('cp -r {0}/* {1}'.format(source_dir, deployment_dir),
          user=deployment_config['user'])
